test, repair truffle test for validator management.
